@@ -1,25 +1,26 @@
 {
-  "nativescript": {
-    "id": "org.nativescript.nativescriptmarketplacedemo",
-    "tns-ios": {
-      "version": "1.3.0"
-    },
-    "tns-android": {
-      "version": "1.4.0"
-    }
-  },
-  "dependencies": {
-    "highlight.js": "^8.7.0",
-    "nativescript-telerik-ui": "file:deps/nativescript-ui/binaries/nativescript-ui-debug-dev.tgz",
-    "settings": "file:settings",
-    "tns-core-modules": "1.3.0"
-  },
-  "devDependencies": {
-    "grunt": "0.4.5",
-    "grunt-ts": "4.2.0",
-    "grunt-contrib-copy": "0.8.0",
-    "grunt-contrib-clean": "0.6.0",
-    "grunt-shell": "1.1.2",
-    "typescript": "1.5.3"
-  }
+	"nativescript": {
+		"id": "org.nativescript.nativescriptmarketplacedemo",
+		"tns-ios": {
+			"version": "1.4.1"
+		},
+		"tns-android": {
+			"version": "1.4.0"
+		}
+	},
+	"dependencies": {
+		"highlight.js": "^8.7.0",
+		"nativescript-telerik-ui": "file:deps/nativescript-ui/binaries/nativescript-ui-stable-dev.tgz",
+		"settings": "file:deps/settings",
+		"tns-core-modules": "^1.3.0"
+	},
+	"devDependencies": {
+        "typescript": "1.6.2",
+		"grunt": "0.4.5",
+        "grunt-ts": "5.0.0-beta.5",
+		"grunt-contrib-copy": "0.8.0",
+		"grunt-contrib-clean": "0.6.0",
+		"grunt-shell": "1.1.2"
+	}
 }
+
