@@ -1,49 +1,3 @@
-/// <reference path="../deps/NativeScript/bin/dist/definitions/declarations.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/es6-promise.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/es-collections.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/platform/platform.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/utils/utils.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/data/observable/observable.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/data/observable-array/observable-array.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/core/view.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/application/application.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/color/color.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/core/dependency-observable.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/core/proxy.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/core/bindable.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/animation/animation.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/content-view/content-view.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/styling/styling.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/styling/style-property.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/styling/css-selector.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/css/reworkcss.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/frame/frame.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/page/page.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/action-bar/action-bar.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/layouts/layout-base.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/layouts/layout.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/layouts/stack-layout/stack-layout.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/layouts/grid-layout/grid-layout.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/layouts/dock-layout/dock-layout.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/builder/builder.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/builder/component-builder.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/gestures/gestures.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/text-base/text-base.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/text/formatted-string.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/text/span.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/editable-text-base/editable-text-base.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/text-view/text-view.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/text-field/text-field.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/button/button.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/repeater/repeater.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/label/label.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/switch/switch.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/image/image.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/image-source/image-source.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/file-system/file-system.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/list-view/list-view.d.ts" />
-/// <reference path="../deps/NativeScript/bin/dist/definitions/ui/dialogs/dialogs.d.ts" />
-
 /// DISABLED reference path="../deps/NativeScript/bin/dist/definitions/android17.d.ts" />
 /// DISABLED reference path="../deps/NativeScript/bin/dist/definitions/ios.d.ts" />
 
@@ -53,3 +7,4 @@
 declare function __startCPUProfiler(string): void;
 declare function __stopCPUProfiler(string): void;
 declare var java: any;
+declare var com: any;
