@@ -11,7 +11,11 @@ import prof = require("../common/profiling");
 
 var page;
 
-export function onPageLoaded(args: observable.EventData) {
+export function pageLoaded(){
+    prof.stop("main-page");
+}
+
+export function onNavigatingTo(args: observable.EventData) {
     // Get the event sender
     page = <pages.Page>args.object;
     page.bindingContext = mainPageVM.instance;
@@ -23,7 +27,6 @@ export function onPageLoaded(args: observable.EventData) {
         bar.translucent = true;
         bar.shadowImage = UIImage.new();
         bar.tintColor = UIColor.whiteColor();
-
 
         // TODO: Is it possible to style the title color of the action bar?
         (<any>bar).titleTextAttributes = { [NSForegroundColorAttributeName]: UIColor.whiteColor() };
@@ -41,7 +44,6 @@ export function toggleWrapLayout(e: any) {
 
 export function navigateToExampleGroup(args: gestures.GestureEventData) {
     prof.start("group");
-
     page.getViewById("side-drawer").closeDrawer();
     var exampleGroup = <examplesVM.ExampleGroup>(<any>args).object.bindingContext;
     var context = new groupPageVM.GroupPageViewModel(exampleGroup);
@@ -49,9 +51,7 @@ export function navigateToExampleGroup(args: gestures.GestureEventData) {
 }
 
 export function navigateToExample(args: gestures.GestureEventData) {
-    //prof.startCPUProfile("example");
     prof.start("example");
-
     page.getViewById("side-drawer").closeDrawer();
     var example = <examplesVM.Example>(<any>args).object.bindingContext;
     navigator.navigateToExample(example);
@@ -62,12 +62,10 @@ export function showSlideout(args) {
 }
 
 export function tapHome(args) {
-    //prof.startCPUProfile("profile");
     page.getViewById("side-drawer").closeDrawer();
 }
 
 export function tapAbout(args) {
-    //prof.stopCPUProfile("profile");
     page.getViewById("side-drawer").closeDrawer();
     navigator.navigateToAbout();
 }
